extra files/requirements.txt
requests>=2.28.0
pandas>=1.5.0
python-dotenv>=0.19.0
alpha-vantage>=2.3.1
beautifulsoup4>=4.11.0
jupyter>=1.0.0 